pyproject.toml
[project]
name = "dqe-agent"
version = "3.0.0"
description = "Browser automation agent with Planner-Executor-Verifier architecture"
requires-python = ">=3.11,<3.14"
dependencies = [
    # --- Orchestration ---
    "langgraph>=0.4,<1.0",
    "langchain>=0.3,<1.0",
    "langchain-openai>=0.3,<1.0",
    "langchain-anthropic>=0.3,<1.0",

    # --- Browser Automation ---
    "playwright>=1.49,<2.0",
    "anthropic>=0.42,<1.0",
    "openai>=1.70,<2.0",

    # --- Data & Config ---
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "python-dotenv>=1.0,<2.0",

    # --- API ---
    "fastapi>=0.115,<1.0",
    "uvicorn>=0.32,<1.0",
    "websockets>=12.0,<15.0",

    # --- Config Files ---
    "pyyaml>=6.0,<7.0",

    # --- Logging ---
    "structlog>=24.0",

    # --- Durable Checkpointing ---
    "langgraph-checkpoint-sqlite>=2.0",
    "aiosqlite>=0.20",

    # --- WebRTC Streaming ---
    "aiortc>=1.9",
    "av>=12.0",
]

[project.optional-dependencies]
mcp = [
    "langchain-mcp-adapters>=0.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "langchain-mcp-adapters>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dqe_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
